PERSIST_INSTALL option followup (#10479)
