+++ changelog/v1.19.0-beta3/k8s-e2e-skip-install-update-followup.yaml
@@ -1,0 +1,6 @@
+changelog:
+  - type: NON_USER_FACING
+    issueLink: https://github.com/solo-io/solo-projects/issues/7432
+    resolvesIssue: true
+    description: >-
+      Cleanup/PR feedback from initial PR.
